Extract main method from Duke to Main class
This is in preparation to integrate JavaFX with Duke, preventing the
Duke class from becoming too bloated.
