+++ build.gradle
@@ -29,7 +29,7 @@ test {
 }
 
 application {
-    mainClass.set("duke.Duke")
+    mainClass.set("duke.Main")
 }
 
 shadowJar {
